Build UI for logging in and creating accounts
