+++ app/src/main/java/alison/fivethingskotlin/CreateAccountActivity.kt
@@ -1,0 +1,12 @@
+package alison.fivethingskotlin
+
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+
+class CreateAccountActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_create_account)
+    }
+}
