+++ app/src/main/java/alison/fivethingskotlin/LoginActivity.kt
@@ -1,0 +1,28 @@
+package alison.fivethingskotlin
+
+import android.content.Context
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import kotlinx.android.synthetic.main.activity_log_in.*
+import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper
+
+class LoginActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_log_in)
+
+        logInButton.setOnClickListener{
+            logIn()
+        }
+    }
+
+    override fun attachBaseContext(newBase: Context) {
+        super.attachBaseContext(CalligraphyContextWrapper.wrap(newBase))
+    }
+
+    private fun logIn() {
+        //TODO
+    }
+
+}
