+++ app/src/main/res/layout/activity_container.xml
@@ -28,7 +28,7 @@
 
             <ImageView
                 android:layout_width="match_parent"
-                android:layout_height="40dp"
+                android:layout_height="35dp"
                 android:layout_gravity="center"
                 android:contentDescription="Five Things Logo"
                 android:src="@drawable/five_things_logo" />
